SSCS-2631 [AS] Refactor controllers to separate functionality.
