+++ src/test/resources/json/sya.json
@@ -1,0 +1,69 @@
+{
+  "benefitType": {
+    "code": "PIP",
+    "description": "Personal Independence Payment"
+  },
+  "postCodeCheck": "B14 7SP",
+  "mrn": {
+    "dwpIssuingOffice": "8",
+    "date": "01-12-2017"
+  },
+  "isAppointee": false,
+  "appellant": {
+    "title": "Mr",
+    "firstName": "Harry",
+    "lastName": "Potter",
+    "dob": "01-05-1990",
+    "nino": "AB877533C",
+    "contactDetails": {
+      "addressLine1": "123 Hairy Lane",
+      "addressLine2": "Off Hairy Park",
+      "townCity": "Hairyfield",
+      "county": "Kent",
+      "postCode": "TN32 6PL",
+      "phoneNumber": "07411222222",
+      "emailAddress": "harry.potter@wizards.com"
+    }
+  },
+  "smsNotify": {
+    "wantsSMSNotifications": true,
+    "useSameNumber": true,
+    "smsNumber": "07411222222"
+  },
+  "hasRepresentative": true,
+  "representative": {
+    "firstName": "Hermione",
+    "lastName": "Granger",
+    "organisation": "Harry Potter Entertainments Ltd",
+    "contactDetails": {
+      "addressLine1": "991 Harlow Road",
+      "addressLine2": "Off Jam Park",
+      "townCity": "Tunbridge Wells",
+      "county": "Kent",
+      "postCode": "TN32 6PL",
+      "phoneNumber": "07411666666",
+      "emailAddress": "hermione.granger@wizards.com"
+    }
+  },
+  "reasonsForAppealing": {
+    "reasons": "Here are my reasons for appealing...",
+    "otherReasons": "Nope, not today anyway!"
+  },
+  "emailAddress": "harry.potter@wizards.com",
+  "hearing": {
+    "wantsToAttend": true,
+    "wantsSupport": true,
+    "arrangements": {
+      "languageInterpreter": true,
+      "signLanguageInterpreter": true,
+      "hearingLoop": true,
+      "disabledAccess": true
+    },
+    "anythingElse": "Nothing else today.",
+    "scheduleHearing": true,
+    "datesCantAttend": [
+      "25/01/1972"
+    ]
+  },
+  "statementOfTruth": true
+}
